fix(dtype): propagate self.dtype through path init and UMA model
Two thin dtype-propagation fixes that surfaced while debugging fp64
chemistry runs:

- popcornn.py: torch.linspace on the initial-path-fit times now passes
  dtype=self.dtype. Without this the linspace defaults to float32 even
  when the Popcornn instance is fp64, then promotes inside initialize_path
  and leaves the optimizer chasing a one-ulp-shifted target.
- uma.py: pass InferenceSettings(base_precision_dtype=self.dtype) into
  fairchem's get_predict_unit. Default is float32; without this an
  fp64 mep ends up with float32 UMA weights and float32 forces despite
  self.dtype=float64 — mixed-precision through autograd that breaks
  fmax_TS quality at the few-meV scale.

+++ popcornn/popcornn.py
@@ -172,7 +172,7 @@ class Popcornn:
         if len(images) > 2:
             self.path = initialize_path(
                 path=self.path, 
-                times=torch.linspace(self.path.t_init.item(), self.path.t_final.item(), len(self.images), device=self.device), 
+                times=torch.linspace(self.path.t_init.item(), self.path.t_final.item(), len(self.images), device=self.device, dtype=self.dtype),
                 init_points=self.images.positions,
             )
